Allow users to configure maximum number of database connections
For both the MySQL and PostgreSQL database backends, allow the user to
configure the maximum number of concurrent database connections.

This is useful in environments where a very limited number of database
connections are available. This is also useful in environments with
strict policies against persistent database connections. In this case,
max_connections can be set to 0, which disables persistent database
connections completely.